RuneRPG - Animation system updated - Target cursor added to battle - Added OnEnter.ini file loading when entering the global state for eased testing. - Removed old file and loads of old code in various files, mostly pertaining the the older InputProcessor approach for handling all types of input. - New RuneBattle and RuneBattleActionLibrary classes added. - Battle.h header added to easily include all files needed for RuneRPG battle coding. - New files GMAnimate added for handling Animations (these were handled in GMSetEntity before, but I feel it might be better to separate them)

+++ src/Maps/Grids/Tile.h
@@ -52,7 +52,8 @@ public:
 	int typeIndex;
 	TileType * type;
 	/// Standard tile position: integral using only X and Y
-	Vector3i position;
+	Vector3f position;
+
 //	const int X() const { return x; };
 //	const int Y() const { return y; };
